Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/navigation-timing/test_unique_performance_objects-expected.txt

Description

This test validates that each window has a unique window.performance object.


PASS window.performance is defined
PASS window.performance.timing is defined
PASS Different window objects have unique performance objects
